Icons should work with the `image(id, src)` API in App Lab
[Fixes #118362451]
