+++ apps/src/applab/designMode.js
@@ -1,4 +1,4 @@
-/* global Applab, dashboard */
+/* global Applab */
 
 // TODO (brent) - make it so that we dont need to specify .jsx. This currently
 // works in our grunt build, but not in tests
@@ -175,26 +175,6 @@ designMode.onPropertyChange = function (element, name, value) {
 };
 
 /**
- * Create a data-URI with the image data of the given icon glyph.
- * @param value {string} An icon identifier of the format "icon://fa-icon-name".
- * @param element {Element}
- * @return {string}
- */
-function renderIconToString(value, element) {
-  var canvas = document.createElement('canvas');
-  canvas.width = canvas.height = 400;
-  var ctx = canvas.getContext('2d');
-  ctx.font = '300px FontAwesome, serif';
-  ctx.textBaseline = 'middle';
-  ctx.textAlign = 'center';
-  ctx.fillStyle = element.getAttribute('data-icon-color');
-  var regex = new RegExp('^' + ICON_PREFIX + 'fa-');
-  var unicode = '0x' + dashboard.iconsUnicode[value.replace(regex, '')];
-  ctx.fillText(String.fromCharCode(unicode), 200, 200);
-  return canvas.toDataURL();
-}
-
-/**
  * After handling properties generically, give elementLibrary a chance
  * to do any element specific changes.
  * @param element
@@ -282,7 +262,7 @@ designMode.updateProperty = function (element, name, value) {
       };
 
       if (ICON_PREFIX_REGEX.test(value)) {
-        element.style.backgroundImage = 'url(' + renderIconToString(value, element) + ')';
+        element.style.backgroundImage = 'url(' + assetPrefix.renderIconToString(value, element) + ')';
         fitImage();
         break;
       }
@@ -305,7 +285,7 @@ designMode.updateProperty = function (element, name, value) {
       var height = parseInt(element.style.height, 10);
       element.style.backgroundSize = width + 'px ' + height + 'px';
 
-      var url = ICON_PREFIX_REGEX.test(value) ? renderIconToString(value, element) : assetPrefix.fixPath(value);
+      var url = ICON_PREFIX_REGEX.test(value) ? assetPrefix.renderIconToString(value, element) : assetPrefix.fixPath(value);
       element.style.backgroundImage = 'url(' + url + ')';
 
       break;
@@ -315,7 +295,7 @@ designMode.updateProperty = function (element, name, value) {
       element.setAttribute('data-canonical-image-url', value);
 
       if (ICON_PREFIX_REGEX.test(value)) {
-        element.src = renderIconToString(value, element);
+        element.src = assetPrefix.renderIconToString(value, element);
         break;
       }
